+++ packages/web/src/lib/myalgowallet-mode.ts
@@ -153,7 +153,7 @@ export class MyAlgoWalletSession {
 	}
 
 	// Function used to wait for a tx confirmation
-	private async waitForConfirmation(
+	async waitForConfirmation(
 		txId: string,
 		waitRounds = WAIT_ROUNDS
 	): Promise<TxnReceipt> {