+++ CHANGELOG.md
@@ -73,7 +73,10 @@ from `algosdk` and sends it to the network.
 
 #### @algo-builder/web
 
-- `executeTx` promise now returns [TxnReceipt](https://github.com/scale-it/algo-builder/blob/master/packages/web/src/types.ts#L458) instead of `algosdk.modelsv2.PendingTransactionResponse`.
+- `executeTx`, `sendAndWait`, `waitForConfirmation` promise now returns [TxnReceipt](https://github.com/scale-it/algo-builder/blob/master/packages/web/src/types.ts#L458) instead of `algosdk.modelsv2.PendingTransactionResponse`.
+
+#### @algo-builder/algob
+- `sendAndWait`, `waitForConfirmation`, `executeSignedTxnFromFile` promise now returns `TxnReceipt` instead of `ConfirmedTxInfo`.
 
 ### Examples